# Conflicts:
#	src/Core/SharedKernels/DataCollection/DataCollection/Implementation/Aggregates/InterviewEntities/InterviewTreeQuestion.cs

+++ src/Services/Export/WB.Services.Export.Host/Controllers/JobController.cs
@@ -1,9 +1,11 @@
 ﻿using System;
+using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using WB.Services.Export.Ddi;
 using WB.Services.Export.Interview;
 using WB.Services.Export.Jobs;
+using WB.Services.Export.Models;
 using WB.Services.Export.Questionnaire;
 using WB.Services.Export.Services.Processing;
 using WB.Services.Infrastructure.Tenant;
@@ -17,47 +19,47 @@ namespace WB.Services.Export.Host.Controllers
     {
         private readonly IDataExportProcessesService exportProcessesService;
         private readonly IJobsStatusReporting jobsStatusReporting;
+        private readonly IExportArchiveHandleService archiveHandleService;
         private readonly IDdiMetadataAccessor ddiDdiMetadataAccessor;
         private readonly IJobService jobService;
 
         public JobController(IDataExportProcessesService exportProcessesService,
             IJobsStatusReporting jobsStatusReporting,
+            IExportArchiveHandleService archiveHandleService,
             IDdiMetadataAccessor ddiDdiMetadataAccessor, IJobService jobService)
         {
             this.exportProcessesService = exportProcessesService;
             this.jobsStatusReporting = jobsStatusReporting;
+            this.archiveHandleService = archiveHandleService;
             this.ddiDdiMetadataAccessor = ddiDdiMetadataAccessor;
             this.jobService = jobService;
         }
 
         [HttpPut]
         [Route("api/v1/job/generate")]
-        public async Task<ActionResult> RequestUpdate(string questionnaireId,
+        public async Task<ActionResult> RequestUpdate(
+            string questionnaireId,
             DataExportFormat format,
             InterviewStatus? status,
             DateTime? from,
             DateTime? to,
-            string archiveName,
             string archivePassword,
             string accessToken,
             ExternalStorageType? storageType,
             TenantInfo tenant)
         {
-            if (string.IsNullOrWhiteSpace(archiveName))
-            {
-                return BadRequest("ArchiveName is required");
-            }
-
             var args = new DataExportProcessArgs
             {
-                Format = format,
-                Questionnaire = new QuestionnaireId(questionnaireId),
-                Tenant = tenant,
-                InterviewStatus = status,
-                FromDate = from,
-                ToDate = to,
+                ExportSettings = new ExportSettings
+                {
+                    Tenant = tenant,
+                    QuestionnaireId = new QuestionnaireId(questionnaireId),
+                    ExportFormat = format,
+                    FromDate = from,
+                    ToDate = to,
+                    Status = status
+                },
                 ArchivePassword = archivePassword,
-                ArchiveName = archiveName,
                 AccessToken = accessToken,
                 StorageType = storageType
             };
@@ -71,16 +73,14 @@ namespace WB.Services.Export.Host.Controllers
         [ResponseCache(NoStore = true)]
         [Route("api/v1/job/status")]
         public async Task<DataExportStatusView> GetDataExportStatusForQuestionnaire(
-            string questionnaireId,
-            string archiveName,
+             string questionnaireId,
             InterviewStatus? status,
             DateTime? fromDate,
             DateTime? toDate,
             TenantInfo tenant)
         {
             var dataExportStatusForQuestionnaire = await this.jobsStatusReporting.GetDataExportStatusForQuestionnaireAsync(tenant,
-                new QuestionnaireId(questionnaireId),
-                archiveName, status, fromDate, toDate);
+                new QuestionnaireId(questionnaireId), status, fromDate, toDate);
 
             return dataExportStatusForQuestionnaire;
         }
@@ -89,11 +89,11 @@ namespace WB.Services.Export.Host.Controllers
         [ResponseCache(NoStore = true)]
         [Route("api/v1/ddi")]
         public async Task<FileStreamResult> GetDdiFile(
-            string questionnaireId,
+            [FromQuery] QuestionnaireId questionnaireId,
             string archivePassword,
             TenantInfo tenant)
         {
-            var pathToFile = await this.ddiDdiMetadataAccessor.GetFilePathToDDIMetadataAsync(tenant, new QuestionnaireId(questionnaireId),
+            var pathToFile = await this.ddiDdiMetadataAccessor.GetFilePathToDDIMetadataAsync(tenant, questionnaireId,
                 archivePassword);
             var responseStream = System.IO.File.OpenRead(pathToFile);
             return File(responseStream, "application/zip");
@@ -102,7 +102,8 @@ namespace WB.Services.Export.Host.Controllers
         [HttpGet]
         [ResponseCache(NoStore = true)]
         [Route("api/v1/job/download")]
-        public async Task<ActionResult> DownloadArchive(string questionnaireId,
+        public async Task<ActionResult> DownloadArchive(
+            string questionnaireId,
             string archiveName,
             InterviewStatus? status,
             DataExportFormat format,
@@ -110,7 +111,17 @@ namespace WB.Services.Export.Host.Controllers
             DateTime? toDate,
             TenantInfo tenant)
         {
-            var result = await this.jobsStatusReporting.DownloadArchiveAsync(tenant, archiveName, format, status, fromDate, toDate);
+            var exportSettings = new ExportSettings
+            {
+                QuestionnaireId = new QuestionnaireId(questionnaireId),
+                ExportFormat = format,
+                Status = status,
+                Tenant = tenant,
+                FromDate = fromDate,
+                ToDate = toDate
+            };
+
+            var result = await this.archiveHandleService.DownloadArchiveAsync(exportSettings, archiveName);
 
             if (result == null)
             {
@@ -119,13 +130,13 @@ namespace WB.Services.Export.Host.Controllers
 
             if (result.Redirect != null)
             {
-                Response.Headers.Add("NewLocation", result.Redirect.ToString());
+                Response.Headers.Add("NewLocation", result.Redirect);
                 return Ok();
             }
 
             return new FileStreamResult(result.Data, "application/octet-stream")
             {
-                FileDownloadName = result.FileName
+                FileDownloadName = WebUtility.UrlEncode(result.FileName)
             };
         }
 
@@ -133,7 +144,7 @@ namespace WB.Services.Export.Host.Controllers
         [Route("api/v1/delete")]
         public async Task<ActionResult> Delete(TenantInfo tenant)
         {
-            await this.jobsStatusReporting.ClearAllExportArchives(tenant);
+            await this.archiveHandleService.ClearAllExportArchives(tenant);
             return Ok();
         }
 
@@ -142,7 +153,11 @@ namespace WB.Services.Export.Host.Controllers
         public async Task<ActionResult> DeleteDataExportProcess(string processId, TenantInfo tenant)
         {
             var job = await jobService.GetJobAsync(tenant, processId, JobStatus.Running, JobStatus.Created);
-            this.exportProcessesService.DeleteDataExport(job.Id, "User canceled");
+            if (job != null)
+            {
+                this.exportProcessesService.DeleteDataExport(job.Id, "User canceled");
+            }
+
             return Ok();
         }
     }
